backend/requirements.txt
# FastAPI and dependencies
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.9
alembic>=1.10.2

# HTTP and networking
httpx>=0.24.0
aiohttp>=3.8.4

# Utilities
python-jose>=3.3.0  # JWT
passlib>=1.7.4      # Password hashing
bcrypt>=4.0.1       # Password hashing
python-multipart>=0.0.6  # Form data parsing

# Data processing
pandas>=2.0.0
numpy>=1.24.2
openpyxl>=3.1.2     # Excel support for pandas

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Development
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0 